+++ primitives/manta/src/assets.rs
@@ -312,6 +312,9 @@ impl FungibleLedgerError {
 ///
 /// [`fungible`]: frame_support::traits::tokens::fungible
 /// [`fungibles`]: frame_support::traits::tokens::fungibles
+///
+/// It is assumed that the supply of native asset cannot be changed,
+/// while the supply of non-native assets can increase or decrease.
 pub trait FungibleLedger<C>
 where
     C: frame_system::Config,
@@ -320,6 +323,7 @@ where
     fn ensure_valid(asset_id: AssetId) -> Result<(), FungibleLedgerError>;
 
     /// Check whether `account` can increase its balance by `amount` in the given `asset_id`.
+    /// Non-native assets will use the `can_increase_total_supply` check, while native assets will not.
     fn can_deposit(
         asset_id: AssetId,
         account: &C::AccountId,
@@ -335,8 +339,9 @@ where
         existence_requirement: ExistenceRequirement,
     ) -> Result<(), FungibleLedgerError>;
 
-    /// Mints `amount` of an asset with the given `asset_id` to `beneficiary`.
-    fn mint(
+    /// Deposit `amount` of an asset with the given `asset_id` to `beneficiary`.
+    /// Will mint and increase the total supply of non-native assets.
+    fn deposit_can_mint(
         asset_id: AssetId,
         beneficiary: &C::AccountId,
         amount: Balance,
@@ -351,8 +356,9 @@ where
         existence_requirement: ExistenceRequirement,
     ) -> Result<(), FungibleLedgerError>;
 
-    /// Performs a burn from `who` for `amount` of `asset_id`
-    fn burn(
+    /// Performs a withdraw from `who` for `amount` of `asset_id`
+    /// Will burn and decrease total supply of non-native assets
+    fn withdraw_can_burn(
         asset_id: AssetId,
         who: &C::AccountId,
         amount: Balance,
@@ -435,17 +441,15 @@ where
     }
 
     #[inline]
-    fn mint(
+    fn deposit_can_mint(
         asset_id: AssetId,
         beneficiary: &C::AccountId,
         amount: Balance,
     ) -> Result<(), FungibleLedgerError> {
         Self::ensure_valid(asset_id)?;
         if asset_id == A::NativeAssetId::get() {
-            Self::can_deposit(asset_id, beneficiary, amount, false)?;
             <Native as Currency<C::AccountId>>::deposit_creating(beneficiary, amount);
         } else {
-            Self::can_deposit(asset_id, beneficiary, amount, true)?;
             <NonNative as Mutate<C::AccountId>>::mint_into(asset_id, beneficiary, amount)
                 .map_err(FungibleLedgerError::InvalidMint)?;
         }
@@ -486,7 +490,7 @@ where
     }
 
     #[inline]
-    fn burn(
+    fn withdraw_can_burn(
         asset_id: AssetId,
         who: &C::AccountId,
         amount: Balance,
@@ -503,8 +507,8 @@ where
             )
             .map_err(FungibleLedgerError::InvalidBurn)?;
         } else {
-            // `existence_requirement` is used in the `can_reduce_by_amount` checks
-            // so it doesn't matter that `burn_from` uses `allow_death` by default in own chosen implementation
+            // `existence_requirement` is used in the `can_reduce_by_amount` checks,
+            // so it doesn't matter that `burn_from` uses `allow_death` by default in our chosen implementation
             <NonNative as Mutate<C::AccountId>>::burn_from(asset_id, who, amount)
                 .map_err(FungibleLedgerError::InvalidBurn)?;
         }